feat: Implement navigation menu and testimonials slider components
